+++ apps/demo/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @authhero/demo
 
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies
+  - authhero@0.23.0
+  - @authhero/kysely-adapter@0.25.0
+
 ## 0.5.0
 
 ### Minor Changes